+++ src/dolphin/workflows/wrapped_phase.py
@@ -54,7 +54,9 @@ def run(cfg: Workflow, debug: bool = False) -> tuple[list[Path], Path, Path]:
     # Make the nodata mask from the polygons, if we're using OPERA CSLCs
     try:
         nodata_mask_file = cfg.scratch_directory / "nodata_mask.tif"
-        _utils.make_nodata_mask(vrt_stack.file_list, out_file=nodata_mask_file)
+        _utils.make_nodata_mask(
+            vrt_stack.file_list, out_file=nodata_mask_file, buffer_pixels=2000
+        )
     except Exception as e:
         logger.warning(f"Could not make nodata mask: {e}")
         nodata_mask_file = None
@@ -99,8 +101,12 @@ def run(cfg: Workflow, debug: bool = False) -> tuple[list[Path], Path, Path]:
         comp_slc_file = next(pl_path.glob("compressed*tif"))
         tcorr_file = next(pl_path.glob("tcorr*tif"))
     else:
-        watcher = NvidiaMemoryWatcher() if utils.gpu_is_available() else None
         logger.info(f"Running sequential EMI step in {pl_path}")
+        if utils.gpu_is_available():  # Track the GPU mem usage if we're using it
+            watcher = NvidiaMemoryWatcher(log_file=pl_path / "nvidia_memory.log")
+        else:
+            watcher = None
+
         if cfg.workflow_name == "single":
             phase_linked_slcs, comp_slc_file, tcorr_file = (
                 single.run_wrapped_phase_single(